requirements.txt
asgiref==3.7.2
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
Django==3.2
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-richtexteditor==0.0.1.dev3
djangorestframework==3.14.0
idna==3.10
pillow==10.3.0
pytz==2023.3.post1
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.3
